+++ hazelcast/src/main/java/com/hazelcast/spi/impl/operationservice/impl/InvocationBuilderImpl.java
@@ -23,30 +23,30 @@ import com.hazelcast.spi.Operation;
 import com.hazelcast.spi.impl.NodeEngineImpl;
 
 /**
- * An {@link com.hazelcast.spi.InvocationBuilder} that is tied to the {@link BasicOperationService}.
+ * An {@link com.hazelcast.spi.InvocationBuilder} that is tied to the {@link OperationServiceImpl}.
  */
-public class BasicInvocationBuilder extends InvocationBuilder {
+public class InvocationBuilderImpl extends InvocationBuilder {
 
-    public BasicInvocationBuilder(NodeEngineImpl nodeEngine, String serviceName, Operation op, int partitionId) {
+    public InvocationBuilderImpl(NodeEngineImpl nodeEngine, String serviceName, Operation op, int partitionId) {
         this(nodeEngine, serviceName, op, partitionId, null);
     }
 
-    public BasicInvocationBuilder(NodeEngineImpl nodeEngine, String serviceName, Operation op, Address target) {
+    public InvocationBuilderImpl(NodeEngineImpl nodeEngine, String serviceName, Operation op, Address target) {
         this(nodeEngine, serviceName, op, Operation.GENERIC_PARTITION_ID, target);
     }
 
-    private BasicInvocationBuilder(NodeEngineImpl nodeEngine, String serviceName, Operation op,
-                                   int partitionId, Address target) {
+    private InvocationBuilderImpl(NodeEngineImpl nodeEngine, String serviceName, Operation op,
+                                  int partitionId, Address target) {
         super(nodeEngine, serviceName, op, partitionId, target);
     }
 
     @Override
     public InternalCompletableFuture invoke() {
         if (target == null) {
-            return new BasicPartitionInvocation(nodeEngine, serviceName, op, partitionId, replicaIndex,
+            return new PartitionInvocation(nodeEngine, serviceName, op, partitionId, replicaIndex,
                     tryCount, tryPauseMillis, callTimeout, callback, resultDeserialized).invoke();
         } else {
-            return new BasicTargetInvocation(nodeEngine, serviceName, op, target, tryCount, tryPauseMillis,
+            return new TargetInvocation(nodeEngine, serviceName, op, target, tryCount, tryPauseMillis,
                     callTimeout, callback, resultDeserialized).invoke();
         }
     }
