+++ hazelcast/src/main/java/com/hazelcast/spi/impl/operationservice/impl/operations/package-info.java
@@ -1,4 +1,4 @@
 /**
- * Contains the Operations being used by the {@link com.hazelcast.spi.impl.operationservice.impl.BasicOperationService}.
+ * Contains the Operations being used by the {@link com.hazelcast.spi.impl.operationservice.impl.OperationServiceImpl}.
  */
 package com.hazelcast.spi.impl.operationservice.impl.operations;
